pyproject.toml
[project]
name = "acadai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.6.2",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "psycopg>=3.3.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
